construir y editar page han sido modificados
