+++ lib/pages/construir/editar_page.dart
@@ -1,24 +1,26 @@
-import 'package:bienaventurados/providers/auth_provider.dart';
-import 'package:bienaventurados/utils/routes.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_icons/flutter_icons.dart';
-import 'package:provider/provider.dart';
 
-class ConstruirPage extends StatefulWidget {
-  final VoidCallback openDrawer;
+class EditarPage extends StatefulWidget {
+  final String? frase;
+  final String? santo;
+  final String? reflexion;
+  final String? tag;
+  final String? usuario;
+  final String? id;
 
-  const ConstruirPage({Key? key, required this.openDrawer}) : super(key: key);
+  const EditarPage({ Key? key, required this.frase, required this.santo, required this.reflexion, required this.tag, required this.usuario, required this.id }) : super(key: key);
 
   @override
-  _ConstruirPageState createState() => _ConstruirPageState();
+  _EditarPageState createState() => _EditarPageState();
 }
 
-class _ConstruirPageState extends State<ConstruirPage> {
+class _EditarPageState extends State<EditarPage> {
   final TextEditingController _fraseController = TextEditingController();
   final TextEditingController _santoController = TextEditingController();
   final TextEditingController _reflexionController = TextEditingController();
   final TextEditingController _tagController = TextEditingController();
+  final TextEditingController _usuarioController = TextEditingController();
   final TextEditingController _preguntaController = TextEditingController();
   final TextEditingController _misionController = TextEditingController();
   final fraseFormKey = GlobalKey<FormState>();
@@ -34,22 +36,20 @@ class _ConstruirPageState extends State<ConstruirPage> {
   }
 
   @override
+  void initState() {
+    _fraseController.text = widget.frase as String;
+    _santoController.text = widget.santo as String;
+    _reflexionController.text = widget.reflexion as String;
+    _tagController.text = widget.tag as String;
+    _usuarioController.text = widget.usuario as String;
+    super.initState();
+  }
+
+  @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.transparent,
+        return Scaffold(
       appBar: AppBar(
-        backgroundColor: Colors.transparent,
         elevation: 0.0,
-        leading: Padding(
-          padding: const EdgeInsets.only(left: 30.0),
-          child: IconButton(
-            onPressed: () {
-              widget.openDrawer();
-            },
-            icon: Icon(FlutterIcons.menu_fea,
-                size: MediaQuery.of(context).size.width * 0.06),
-          ),
-        ),
       ),
       body: PageView(
         physics: NeverScrollableScrollPhysics(),
@@ -57,33 +57,17 @@ class _ConstruirPageState extends State<ConstruirPage> {
         children: [
           fraseWidget(),
           reflexionWidget(),
-          // preguntaWidget(),
+          preguntaWidget(),
           avioncitoListoWidget(),
         ],
       ),
     );
   }
 
-  void _construirNuevoAvioncito(String? usuario) {
-    final FirebaseFirestore _db = FirebaseFirestore.instance;
-    DocumentReference avioncitoRef = _db.collection('datosUsuarios').doc();
-
-    avioncitoRef.set({
-      'frase': _fraseController.text,
-      'santo': _santoController.text,
-      'reflexion': _reflexionController.text,
-      'tag': _tagController.text,
-      'pregunta': _preguntaController.text,
-      'mision': _misionController.text,
-      'usuario': usuario,
-    });
-  }
-
   Widget fraseWidget() {
     return Form(
       key: fraseFormKey,
       child: Scaffold(
-        backgroundColor: Colors.transparent,
         body: SingleChildScrollView(
           child: Padding(
             padding: const EdgeInsets.all(30.0),
@@ -197,7 +181,6 @@ class _ConstruirPageState extends State<ConstruirPage> {
   }
 
   Widget reflexionWidget() {
-    final authProvider = Provider.of<AuthProvider>(context);
     return Form(
       key: reflexionFormKey,
       child: Scaffold(
@@ -246,40 +229,40 @@ class _ConstruirPageState extends State<ConstruirPage> {
                     }
                   },
                 ),
-                // SizedBox(
-                //   height: MediaQuery.of(context).size.width * 0.04,
-                // ),
-                // TextFormField(
-                //   controller: _tagController,
-                //   autofocus: false,
-                //   keyboardType: TextInputType.text,
-                //   style: Theme.of(context).textTheme.headline2!.copyWith(
-                //         fontSize: MediaQuery.of(context).size.width * 0.06,
-                //       ),
-                //   cursorColor: Theme.of(context).primaryColorDark,
-                //   cursorWidth: 4,
-                //   decoration: InputDecoration(
-                //     hintText: 'Tag',
-                //     hintStyle: Theme.of(context).textTheme.headline2!.copyWith(
-                //         fontSize: MediaQuery.of(context).size.width * 0.06,
-                //         color: Theme.of(context)
-                //             .primaryColorDark
-                //             .withOpacity(0.2)),
-                //     focusedBorder: InputBorder.none,
-                //     border: InputBorder.none,
-                //     errorBorder: InputBorder.none,
-                //     enabledBorder: InputBorder.none,
-                //     disabledBorder: InputBorder.none,
-                //     focusedErrorBorder: InputBorder.none,
-                //   ),
-                //   validator: (value) {
-                //     if (value!.isEmpty) {
-                //       return 'Debes ingresar un nombre para continuar';
-                //     } else {
-                //       return null;
-                //     }
-                //   },
-                // ),
+                SizedBox(
+                  height: MediaQuery.of(context).size.width * 0.04,
+                ),
+                TextFormField(
+                  controller: _tagController,
+                  autofocus: false,
+                  keyboardType: TextInputType.text,
+                  style: Theme.of(context).textTheme.headline2!.copyWith(
+                        fontSize: MediaQuery.of(context).size.width * 0.06,
+                      ),
+                  cursorColor: Theme.of(context).primaryColorDark,
+                  cursorWidth: 4,
+                  decoration: InputDecoration(
+                    hintText: 'Tag',
+                    hintStyle: Theme.of(context).textTheme.headline2!.copyWith(
+                        fontSize: MediaQuery.of(context).size.width * 0.06,
+                        color: Theme.of(context)
+                            .primaryColorDark
+                            .withOpacity(0.2)),
+                    focusedBorder: InputBorder.none,
+                    border: InputBorder.none,
+                    errorBorder: InputBorder.none,
+                    enabledBorder: InputBorder.none,
+                    disabledBorder: InputBorder.none,
+                    focusedErrorBorder: InputBorder.none,
+                  ),
+                  validator: (value) {
+                    if (value!.isEmpty) {
+                      return 'Debes ingresar un nombre para continuar';
+                    } else {
+                      return null;
+                    }
+                  },
+                ),
               ],
             ),
           ),
@@ -295,26 +278,10 @@ class _ConstruirPageState extends State<ConstruirPage> {
               ),
               onPressed: () {
                 if (reflexionFormKey.currentState!.validate()) {
-                  // _pageController.nextPage(
-                  //     duration: Duration(milliseconds: 1200),
-                  //     curve: Curves.fastLinearToSlowEaseIn);
-                  FocusScopeNode currentFocus = FocusScope.of(context);
-
-                  if (!currentFocus.hasPrimaryFocus) {
-                    currentFocus.unfocus();
-                  }
-                  setState(() {
-                    _construirNuevoAvioncito(authProvider.usuario.nombre);
-                    _fraseController.clear();
-                    _santoController.clear();
-                    _reflexionController.clear();
-                    _tagController.clear();
-                    _preguntaController.clear();
-                    _misionController.clear();
-                  });
                   _pageController.nextPage(
                       duration: Duration(milliseconds: 1200),
                       curve: Curves.fastLinearToSlowEaseIn);
+
                 }
               },
               child: Text('Siguiente',
@@ -329,7 +296,6 @@ class _ConstruirPageState extends State<ConstruirPage> {
   }
 
   Widget preguntaWidget() {
-    //final authProvider = Provider.of<AuthProvider>(context);
     return Form(
       key: preguntaFormKey,
       child: Scaffold(
@@ -410,25 +376,26 @@ class _ConstruirPageState extends State<ConstruirPage> {
                 backgroundColor: Theme.of(context).primaryColorDark,
               ),
               onPressed: () {
-                // if (preguntaFormKey.currentState!.validate()) {
-                //   FocusScopeNode currentFocus = FocusScope.of(context);
+                if (preguntaFormKey.currentState!.validate()) {
+                  FocusScopeNode currentFocus = FocusScope.of(context);
 
-                //   if (!currentFocus.hasPrimaryFocus) {
-                //     currentFocus.unfocus();
-                //   }
-                //   setState(() {
-                //     _construirNuevoAvioncito(authProvider.usuario.nombre);
-                //     _fraseController.clear();
-                //     _santoController.clear();
-                //     _reflexionController.clear();
-                //     _tagController.clear();
-                //     _preguntaController.clear();
-                //     _misionController.clear();
-                //   });
-                //   _pageController.nextPage(
-                //       duration: Duration(milliseconds: 1200),
-                //       curve: Curves.fastLinearToSlowEaseIn);
-                // }
+                  if (!currentFocus.hasPrimaryFocus) {
+                    currentFocus.unfocus();
+                  }
+                  setState(() {
+                    _construirNuevoAvioncito();
+                    _deleteAvioncitoUsuario(widget.id);
+                    _fraseController.clear();
+                    _santoController.clear();
+                    _reflexionController.clear();
+                    _tagController.clear();
+                    _preguntaController.clear();
+                    _misionController.clear();
+                  });
+                  _pageController.nextPage(
+                      duration: Duration(milliseconds: 1200),
+                      curve: Curves.fastLinearToSlowEaseIn);
+                }
               },
               child: Text('Continuar',
                   style: Theme.of(context).textTheme.headline4!.copyWith(
@@ -456,7 +423,7 @@ class _ConstruirPageState extends State<ConstruirPage> {
         Padding(
           padding: const EdgeInsets.only(left: 40.0, right: 40.0, top: 40.0),
           child: Text(
-              'Tu avioncito está terminado y pronto empezará a volar por los corazones de todos en Bienaventurados.\n\n¡Muchas gracias por construir!',
+              'Este avioncito ya salió volando y llegará pronto a muchos corazones!',
               style: Theme.of(context).textTheme.bodyText1!.copyWith(
                     fontSize: MediaQuery.of(context).size.width * 0.04,
                   )),
@@ -472,7 +439,7 @@ class _ConstruirPageState extends State<ConstruirPage> {
               backgroundColor: Theme.of(context).primaryColorDark,
             ),
             onPressed: () {
-              Navigator.of(context).pushNamed(dashboardPage);
+              Navigator.of(context).pop();
             },
             child: Text('Finalizar',
                 style: Theme.of(context).textTheme.headline4!.copyWith(
@@ -482,5 +449,230 @@ class _ConstruirPageState extends State<ConstruirPage> {
         ),
       ),
     );
+  }
+
+  //   return Scaffold(
+  //     appBar: AppBar(
+  //       title: Text('Taller de avioncitos',
+  //           style: Theme.of(context).textTheme.headline6),
+  //       elevation: 0.0,
+  //     ),
+  //     body: SingleChildScrollView(
+  //       child: Column(
+  //         children: [
+  //           Padding(
+  //             padding: const EdgeInsets.all(40.0),
+  //             child: Text('Avioncito en construcción...',
+  //                 style: Theme.of(context).textTheme.headline1),
+  //           ),
+  //           Padding(
+  //             padding:
+  //                 const EdgeInsets.only(left: 40.0, right: 40.0, top: 10.0),
+  //             child: TextFormField(
+  //               controller: _fraseController,
+  //               autofocus: false,
+  //               keyboardType: TextInputType.text,
+  //               style: Theme.of(context).textTheme.headline1,
+  //               cursorColor: Theme.of(context).primaryColorDark,
+  //               cursorWidth: 4,
+  //               decoration: InputDecoration(
+  //                 hintText: 'Frase',
+  //                 hintStyle: Theme.of(context).textTheme.headline1!.copyWith(
+  //                     color:
+  //                         Theme.of(context).primaryColorDark.withOpacity(0.2)),
+  //                 focusedBorder: InputBorder.none,
+  //                 border: InputBorder.none,
+  //                 errorBorder: InputBorder.none,
+  //                 enabledBorder: InputBorder.none,
+  //                 disabledBorder: InputBorder.none,
+  //                 focusedErrorBorder: InputBorder.none,
+  //               ),
+  //               validator: (value) {
+  //                 if (value!.isEmpty) {
+  //                   return 'Debes ingresar un nombre para continuar';
+  //                 } else {
+  //                   return null;
+  //                 }
+  //               },
+  //             ),
+  //           ),
+  //           Padding(
+  //             padding:
+  //                 const EdgeInsets.only(left: 40.0, right: 40.0, top: 10.0),
+  //             child: TextFormField(
+  //               controller: _santoController,
+  //               autofocus: false,
+  //               keyboardType: TextInputType.text,
+  //               style: Theme.of(context).textTheme.headline1,
+  //               cursorColor: Theme.of(context).primaryColorDark,
+  //               cursorWidth: 4,
+  //               decoration: InputDecoration(
+  //                 hintText: 'Santo',
+  //                 hintStyle: Theme.of(context).textTheme.headline1!.copyWith(
+  //                     color:
+  //                         Theme.of(context).primaryColorDark.withOpacity(0.2)),
+  //                 focusedBorder: InputBorder.none,
+  //                 border: InputBorder.none,
+  //                 errorBorder: InputBorder.none,
+  //                 enabledBorder: InputBorder.none,
+  //                 disabledBorder: InputBorder.none,
+  //                 focusedErrorBorder: InputBorder.none,
+  //               ),
+  //               validator: (value) {
+  //                 if (value!.isEmpty) {
+  //                   return 'Debes ingresar un nombre para continuar';
+  //                 } else {
+  //                   return null;
+  //                 }
+  //               },
+  //             ),
+              
+  //           ),
+  //           Padding(
+  //             padding:
+  //                 const EdgeInsets.only(left: 40.0, right: 40.0, top: 10.0),
+  //             child: TextFormField(
+  //               controller: _reflexionController,
+  //               autofocus: false,
+  //               keyboardType: TextInputType.text,
+  //               style: Theme.of(context).textTheme.headline1,
+  //               cursorColor: Theme.of(context).primaryColorDark,
+  //               cursorWidth: 4,
+  //               minLines: 4,
+  //               maxLines: 8,
+  //               maxLength: 200,
+  //               decoration: InputDecoration(
+  //                 hintText: 'Reflexión',
+  //                 hintStyle: Theme.of(context).textTheme.headline1!.copyWith(
+  //                     color:
+  //                         Theme.of(context).primaryColorDark.withOpacity(0.2)),
+  //                 focusedBorder: InputBorder.none,
+  //                 border: InputBorder.none,
+  //                 errorBorder: InputBorder.none,
+  //                 enabledBorder: InputBorder.none,
+  //                 disabledBorder: InputBorder.none,
+  //                 focusedErrorBorder: InputBorder.none,
+  //               ),
+  //               validator: (value) {
+  //                 if (value!.isEmpty) {
+  //                   return 'Debes ingresar un nombre para continuar';
+  //                 } else {
+  //                   return null;
+  //                 }
+  //               },
+  //             ),),
+  //             Padding(
+  //             padding:
+  //                 const EdgeInsets.only(left: 40.0, right: 40.0, top: 10.0),
+  //             child: TextFormField(
+  //               controller: _tagController,
+  //               autofocus: false,
+  //               keyboardType: TextInputType.text,
+  //               style: Theme.of(context).textTheme.headline1,
+  //               cursorColor: Theme.of(context).primaryColorDark,
+  //               cursorWidth: 4,
+  //               decoration: InputDecoration(
+  //                 hintText: 'Tag',
+  //                 hintStyle: Theme.of(context).textTheme.headline1!.copyWith(
+  //                     color:
+  //                         Theme.of(context).primaryColorDark.withOpacity(0.2)),
+  //                 focusedBorder: InputBorder.none,
+  //                 border: InputBorder.none,
+  //                 errorBorder: InputBorder.none,
+  //                 enabledBorder: InputBorder.none,
+  //                 disabledBorder: InputBorder.none,
+  //                 focusedErrorBorder: InputBorder.none,
+  //               ),
+  //               validator: (value) {
+  //                 if (value!.isEmpty) {
+  //                   return 'Debes ingresar un nombre para continuar';
+  //                 } else {
+  //                   return null;
+  //                 }
+  //               },
+  //             ),),
+  //             Padding(
+  //             padding:
+  //                 const EdgeInsets.only(left: 40.0, right: 40.0, top: 10.0),
+  //             child: TextFormField(
+  //               controller: _usuarioController,
+  //               autofocus: false,
+  //               keyboardType: TextInputType.text,
+  //               style: Theme.of(context).textTheme.headline1,
+  //               cursorColor: Theme.of(context).primaryColorDark,
+  //               cursorWidth: 4,
+  //               decoration: InputDecoration(
+  //                 hintText: 'Usuario',
+  //                 hintStyle: Theme.of(context).textTheme.headline1!.copyWith(
+  //                     color:
+  //                         Theme.of(context).primaryColorDark.withOpacity(0.2)),
+  //                 focusedBorder: InputBorder.none,
+  //                 border: InputBorder.none,
+  //                 errorBorder: InputBorder.none,
+  //                 enabledBorder: InputBorder.none,
+  //                 disabledBorder: InputBorder.none,
+  //                 focusedErrorBorder: InputBorder.none,
+  //               ),
+  //               validator: (value) {
+  //                 if (value!.isEmpty) {
+  //                   return 'Debes ingresar un nombre para continuar';
+  //                 } else {
+  //                   return null;
+  //                 }
+  //               },
+  //             ),
+              
+  //           ),
+  //           SizedBox(height: 60),
+  //           Padding(
+  //             padding: const EdgeInsets.all(40.0),
+  //             child: InkWell(
+  //                 onTap: () async {
+  //                   setState(() {
+  //                         // _deleteAvioncitoUsuario(widget.id);
+  //                         _deleteAvioncitoUsuario(widget.id);
+  //                         _construirNuevoAvioncito();
+  //                         _fraseController.clear();
+  //                         _santoController.clear();
+  //                         _reflexionController.clear();
+  //                         _tagController.clear();
+  //                         _usuarioController.clear();
+  //                         Navigator.of(context).pop();
+  //                       });
+  //                 },
+  //                 child: Text(
+  //                   'Confirmar',
+  //                   style: Theme.of(context)
+  //                       .textTheme
+  //                       .headline1!
+  //                       .copyWith(color: Theme.of(context).colorScheme.secondary),
+  //                 ),
+  //               )
+  //               ),
+            
+  //         ],
+  //       ),
+  //     ),
+  //   );
+  // }
+
+  void _construirNuevoAvioncito() {
+    final FirebaseFirestore _db = FirebaseFirestore.instance;
+    DocumentReference avioncitoRef = _db.collection('datosApp').doc();
+
+    avioncitoRef.set({
+      'frase': _fraseController.text,
+      'santo': _santoController.text,
+      'reflexion': _reflexionController.text,
+      'tag': _tagController.text,
+      'usuario': _usuarioController.text,
+    });
+  }
+
+  void _deleteAvioncitoUsuario(String? id) {
+    final FirebaseFirestore _db = FirebaseFirestore.instance;
+    DocumentReference avioncitoRef = _db.collection('datosUsuarios').doc(id);
+    print('avioncito ${avioncitoRef.id} eliminado de datosUsuarios');
+    avioncitoRef.delete();
   }
 }
